+++ packages/drizzle/src/index.ts
@@ -1,2 +1,0 @@
-export * from './generateDrizzleRuleGroupProcessor';
-export * from './generateDrizzleRuleProcessor';