+++ packages/drizzle/vitest.config.ts
@@ -1,9 +1,0 @@
-import { mergeConfig } from 'vitest/config';
-import shared from '../../vitest.shared';
-
-export default mergeConfig(shared, {
-  test: {
-    environment: 'node',
-    exclude: ['**/drizzle.sqlite.test.ts'],
-  },
-});